+++ src/App.tsx
@@ -3,9 +3,9 @@ import { RouterProvider } from 'react-router-dom';
 
 import './App.css';
 
-import { authService } from './services/AuthService';
 import { API_FAILED } from './constants/api';
 import { router } from './routes/router';
+import { authService, memberService } from './services';
 
 function App() {
   const [isInit, setIsInit] = useState<boolean>(false);
@@ -15,7 +15,7 @@ function App() {
       .refreshToken()
       .then((res) => {
         if (res.status === API_FAILED) return;
-        return authService.getUserData(res.data.access_token, {
+        return memberService.getUserData(res.data.access_token, {
           headers: {
             Authorization: `Bearer ${res.data.access_token}`,
           },
