+++ src/components/common/Header.tsx
@@ -1,11 +1,12 @@
 import { useEffect, useState } from 'react';
 import { Link, useLocation, useNavigate } from 'react-router-dom';
+import { FaBell, FaCircle } from 'react-icons/fa6';
 import { twMerge } from 'tailwind-merge';
 import { motion } from 'framer-motion';
-import { FaBell, FaCircle } from 'react-icons/fa6';
+
 import { useAuthStore } from '@/store/useAuthStore';
-import { authService } from '@/services/AuthService';
 import { API_FAILED } from '@/constants/api';
+import { authService } from '@/services';
 import Button from './Button';
 
 type InitialIsShowState = {
