+++ src/types/api/member.ts
@@ -1,0 +1,5 @@
+export type GetUserDataResponse = {
+  memberId: number;
+  memberName: string;
+  isCompleteProfile: boolean;
+};
